# SSE2 str_tolower