use sse2 instructions to speed up zend_str_tolower