@nikic 2020-04-27
Any perf numbers? Maybe there is a measurable improvement on something like dynamic function call...