@laruence 2020-04-27
according to : https://gist.github.com/laruence/df51017db658ee2c434e98574615b7d2
without sse2
````
calling A using 0.11974000930786s
calling Aaaa using 0.1306459903717s
calling Aaaaaaaa using 0.14504194259644s
calling Aaaaaaaaaaaaaaaa using 0.18059492111206s
````
with sse2
````
calling A using 0.1224319934845s
calling Aaaa using 0.12959408760071s
calling Aaaaaaaa using 0.14206290245056s
calling Aaaaaaaaaaaaaaaa using 0.14770197868347s
````
there is visible speedup while function name longer than 16.